% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.masreml}
\alias{as.data.frame.masreml}
\title{Convert masreml result to data.frame of GEBVs}
\usage{
\method{as.data.frame}{masreml}(x, include_components = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{"masreml"}}

\item{include_components}{logical, include per-component EBV columns}
}
\value{
data.frame
}
\description{
Convert masreml result to data.frame of GEBVs
}
