% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_positive_definite}
\alias{is_positive_definite}
\title{Check if matrix is positive definite}
\usage{
is_positive_definite(G, label = "G")
}
\arguments{
\item{G}{numeric matrix}

\item{label}{character, name for error messages}
}
\value{
logical
}
\description{
Check if matrix is positive definite
}
