% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jitter_diagonal}
\alias{jitter_diagonal}
\title{Jitter G matrix diagonal to ensure positive definiteness}
\usage{
jitter_diagonal(G, eps = 1e-06)
}
\arguments{
\item{G}{numeric matrix}

\item{eps}{numeric, small value added to diagonal (default 1e-6)}
}
\value{
numeric matrix
}
\description{
Jitter G matrix diagonal to ensure positive definiteness
}
