% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot.masreml}
\alias{plot.masreml}
\title{Plot method for masreml objects}
\usage{
\method{plot}{masreml}(x, type = "gebv", ...)
}
\arguments{
\item{x}{object of class \code{"masreml"}}

\item{type}{character: "gebv" (default), "varcomp", "both"}

\item{...}{additional arguments passed to plot}
}
\description{
Plot method for masreml objects
}
