% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot.masreml_cv}
\alias{plot.masreml_cv}
\title{Plot method for masreml_cv objects}
\usage{
\method{plot}{masreml_cv}(x, ...)
}
\arguments{
\item{x}{object of class \code{"masreml_cv"}}

\item{...}{additional arguments passed to plot}
}
\description{
Plot method for masreml_cv objects
}
