% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.masreml}
\alias{print.masreml}
\title{Print method for masreml objects}
\usage{
\method{print}{masreml}(x, ...)
}
\arguments{
\item{x}{object of class \code{"masreml"}}

\item{...}{ignored}
}
\description{
Print method for masreml objects
}
