% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.masreml_cv}
\alias{print.masreml_cv}
\title{Print method for masreml_cv objects}
\usage{
\method{print}{masreml_cv}(x, ...)
}
\arguments{
\item{x}{object of class \code{"masreml_cv"}}

\item{...}{ignored}
}
\description{
Print method for masreml_cv objects
}
