% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_masreml_threads}
\alias{set_masreml_threads}
\title{Set global number of threads for masreml}
\usage{
set_masreml_threads(n)
}
\arguments{
\item{n}{integer, number of threads}
}
\description{
Set global number of threads for masreml
}
