% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_A_ped}
\alias{build_A_ped}
\title{Utility functions for masreml}
\usage{
build_A_ped(pedigree)
}
\arguments{
\item{pedigree}{data.frame with columns:
\code{id}, \code{sire}, \code{dam}
IDs must be integers or convertible to integer (1-based).
Use 0 for unknown parents.}
}
\value{
numeric matrix (n x n), additive relationship matrix A
}
\description{
Utility functions for masreml
}
\examples{
\dontrun{
ped <- data.frame(
  id   = 1:5,
  sire = c(0, 0, 1, 1, 3),
  dam  = c(0, 0, 2, 2, 4)
)
A <- build_A_ped(ped)
}

}
\keyword{(Henderson}
\keyword{1976)}
\keyword{A}
\keyword{Build}
\keyword{internal}
\keyword{matrix}
\keyword{pedigree}
