% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_D_snp}
\alias{build_D_snp}
\title{Build SNP dominance D matrix (Da et al. 2014)}
\usage{
build_D_snp(W)
}
\arguments{
\item{W}{numeric matrix (n x m), raw genotype 0/1/2}
}
\value{
numeric matrix (n x n), SNP dominance D
}
\description{
Build SNP dominance D matrix (Da et al. 2014)
}
