% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_G_mh}
\alias{build_G_mh}
\title{Build MH additive Agh matrix (Da 2015)}
\usage{
build_G_mh(mh_list)
}
\arguments{
\item{mh_list}{list of data.frames, one per chromosome.
Each data.frame: first col = ID, then paired strand columns
per haplotype block (strand1, strand2, strand1, strand2, ...)}
}
\value{
numeric matrix (n x n), MH additive Agh
}
\description{
Build MH additive Agh matrix (Da 2015)
}
