% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_G_snp}
\alias{build_G_snp}
\title{Build SNP additive G matrix (VanRaden 2008)}
\usage{
build_G_snp(W)
}
\arguments{
\item{W}{numeric matrix (n x m), raw genotype 0/1/2}
}
\value{
numeric matrix (n x n), SNP additive G
}
\description{
Build SNP additive G matrix (VanRaden 2008)
}
