% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.masreml}
\alias{summary.masreml}
\title{Summary method for masreml objects}
\usage{
\method{summary}{masreml}(object, ...)
}
\arguments{
\item{object}{object of class \code{"masreml"}}

\item{...}{ignored}
}
\description{
Summary method for masreml objects
}
