% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.masreml_cv}
\alias{summary.masreml_cv}
\title{Summary method for masreml_cv objects}
\usage{
\method{summary}{masreml_cv}(object, ...)
}
\arguments{
\item{object}{object of class \code{"masreml_cv"}}

\item{...}{ignored}
}
\description{
Summary method for masreml_cv objects
}
