% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{varcomp}
\alias{varcomp}
\title{Extract variance components as data.frame}
\usage{
varcomp(x)
}
\arguments{
\item{x}{object of class \code{"masreml"}}
}
\value{
data.frame with Component, Sigma2, H2, Proportion
}
\description{
Extract variance components as data.frame
}
